MPS-16788 Joda time library should be replaced with the version - 2.1

+++ languages/util/dates/runtime/source_gen/jetbrains/mps/baseLanguage/dates/runtime/DateTimeOperations.java
@@ -363,7 +363,7 @@ public class DateTimeOperations {
     if (leftExpression == null || rightExpression == null) {
       result = null;
     } else {
-      result = new Period(rightExpression, leftExpression, periodType);
+      result = new Period((long) rightExpression, (long) leftExpression, periodType);
     }
     return result;
   }
